\name{gapminder}
\alias{gapminder}
\title{Creates a Gapminder data source (S3 object).}
\description{Creates a Gapminder data source (S3 object).}
\details{Gapminder describes itself as a "fact tank" that promotes a fact based world view. 
On their website they provide a service that allows to create animated charts for various indicators, differentiated by country.
They also provide the underlying datasets for download. This S3 class serves as a wrapper for easy access to a subset of these data.

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\value{an object of class "gapminder"}
\references{\url{http://www.gapminder.org}}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

