\name{gmBirthsAttendedBySkilledHealthStaff-data}
\alias{gmBirthsAttendedBySkilledHealthStaff-data}
\alias{gmBirthsAttendedBySkilledHealthStaff-data}
\title{Births attended by skilled health staff (\% of total)...}
\description{Births attended by skilled health staff (\% of total)}
\details{Births attended by skilled health staff (\% of total)
Definition and explanations
Indicator name Births attended by skilled health staff (\% of total)
Definition of indicator The percentage of deliveries attended by personnel (including doctors,
nurses and midwives) trained to give the necessary supervision, care, and advice to women during pregnancy,
labor, and the postpartum period; to conduct deliveries on their own; and to care for newborns.
Unit of measurement
Data source
Source organization(s) World Bank
Link to source organization http://www.worldbank.org
Complete reference World Development Indicators Online. 2008.
Link to complete reference http://go.worldbank.org/U0FSM7AQ40
Specific information about this indicator
Uploader Gapminder
Date uploaded 6/4/2008 14:48

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmBirthsAttendedBySkilledHealthStaff}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=pyj6tScZqmEfKY9bk02DBYA&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

