\name{gmBroadbandSubscribers-data}
\alias{gmBroadbandSubscribers-data}
\alias{gmBroadbandSubscribers-data}
\title{Broadband subscribers...}
\description{Broadband subscribers}
\details{Broadband subscribers
Definition and explanations
Indicator name Broadband subscribers
Definition of indicator Broadband subscribers are the total number of broadband subscribers with a
digital subscriber line, cable modem, or other high-speed technologies, although definitions may vary from
country to country.
Unit of measurement
Data source
Source organization(s) World Bank
Link to source organization http://www.worldbank.org
Complete reference World Development Indicators Online. 2008.
Link to complete reference http://go.worldbank.org/U0FSM7AQ40
Specific information about this indicator
Uploader Gapminder
Date uploaded 6/4/2008 14:48

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmBroadbandSubscribers}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=pyj6tScZqmEcuy6dYkzGhfw&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

