\name{gmCellPhones-data}
\alias{gmCellPhones-data}
\alias{gmCellPhones-data}
\title{Cell phones (total)...}
\description{Cell phones (total)}
\details{Cell phones (total)
Definition and explanations
Indicator name Cell phones (total)
Definition of indicator The number of subscribers to a service providing access to the telephone
network using cellular technology, including post-paid and pre-paid cell phones.
Unit of measurement
Data source 
Source organization(s) UN data
Link to source organization http://data.un.org/
Complete reference
Link to complete reference
Specific information about this indicator
Uploader Gapminder
[Add other fields as required]

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmCellPhones}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=pyj6tScZqmEcKuNdFCUo6TQ&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

