\name{gmCO2Emissions-data}
\alias{gmCO2Emissions-data}
\alias{gmCO2Emissions-data}
\title{Total CO2 emissions from fossil-fuels (metric tons)...}
\description{Total CO2 emissions from fossil-fuels (metric tons)}
\details{Total CO2 emissions from fossil-fuels (metric tons)
Definition and explanations
Indicator name Total CO2 emissions from fossil-fuels (metric tons)
Definition of indicator Total amount of CO2 in metric tons emitted from the consumption of all kinds of
fossil-fuels.
Unit of measurement
Data source 
Source organization(s) CDIAC (Carbon Dioxide Information Analysis Center)
Link to source organization http://cdiac.ornl.gov/
Complete reference CDIAC (Carbon Dioxide Information Analysis Center) - nation.1751_2006.csv
Link to complete reference http://cdiac.ornl.gov/ftp/ndp030/CSV-FILES/
Specific information about this indicator
Uploader Gapminder

Calculation method Original data was only the amount of Carbon (C) and calculation has been done to
convert Carbon into Carbon Dioxide (CO2): (12+16*2)/12.

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmCO2Emissions}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=phAwcNAVuyj1NHPC9MyZ9SQ&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

