\name{gmCrudeBirthRate-data}
\alias{gmCrudeBirthRate-data}
\alias{gmCrudeBirthRate-data}
\title{Crude birth rate (births per 1000 population)...}
\description{Crude birth rate (births per 1000 population)}
\details{Crude birth rate (births per 1000 population)
Definition and explanations
Indicator name Crude birth rate (births per 1000 population)
Definition of indicator Crude birth rate (births per 1000 population)
Unit of measurement
Data source 
Source organization(s) UN Population Division
Link to source organization http://esa.un.org/peps/index.htm
Complete reference
Link to complete reference
Specific information about this indicator
Uploader Gapminder
[Add other fields as required]

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmCrudeBirthRate}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=tUSeGJOQhafugwUvHvY-wLA&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

