\name{gmCrudeDeathRate-data}
\alias{gmCrudeDeathRate-data}
\alias{gmCrudeDeathRate-data}
\title{Crude death rate (deaths per 1,000 population)...}
\description{Crude death rate (deaths per 1,000 population)}
\details{Crude death rate (deaths per 1,000 population)
Definition and explanations
Indicator name Crude death rate (deaths per 1,000 population)
Definition of indicator Crude death rate (deaths per 1,000 population)
Unit of measurement
Data source 
Source organization(s) UN Population Division
Link to source organization http://esa.un.org/peps/index.htm
Complete reference
Link to complete reference
Specific information about this indicator
Uploader Gapminder
[Add other fields as required]

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmCrudeDeathRate}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=tHyj-2jRvK3CCNJOc5Vm-HQ&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

