\name{gmIncomePerCapita-data}
\alias{gmIncomePerCapita-data}
\alias{gmIncomePerCapita-data}
\title{Income per person (fixed PPP$)...}
\description{Income per person (fixed PPP$)}
\details{Gapminder has compiled the data you see in this graph from several sources, such as official
international statistics, various historical sources and our own estimates. 
The link below takes you to
Gapminder's documentation page, which contains the details on how the compilation was done and the sources for each
observation. 

Definition and explanations
Definition of indicator Gross Domestic Product per capita by Purchasing Power Parities (in
international dollars, fixed 2005 prices). The inflation and differences in the cost of living between countries has
been taken into account.
Main sources: Cross-country data for 2005 is mainly based on the 2005 round of the International
Comparison Program. Estimates based on other sources were used for the other countries. Real growth rates were
linked to the 2005 levels. Several sources are used for these growth rates, such as the data of Angus Maddison.
Follow the link below to download the detailed documentation.
Who has compiled the data and how was it done
Compiled by (i.e. who has compiled the data in the indicator) Gapminder
Link to homepage, if any, of this compiler www.gapminder.org
Link to the full documentation of this indicator
www.gapminder.org/downloads/documentation/#gd001
Updates
12/4/08 Data for Iraq & Zimbabwe added upto 2007
12/10/08 Data for many other countries added upto 2007
12/22/08 Complete revision of the whole data set. This new version is based on the PPPs from the 2005
round from ICP. Previous version of the data set is available as a separate indicator.
4/27/09 Version 5: updated Swedish data, revised Myanmar, raised lowest level, added data from Barro &
Ursua
10/1/09 Version 6: Congo Dem. Rep. revised
12/18/09 Version 7: data for 2008 added

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmIncomePerCapita}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=phAwcNAVuyj1jiMAkmq1iMg&gid=2&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

