\name{gmInfantMortalityRate-data}
\alias{gmInfantMortalityRate-data}
\alias{gmInfantMortalityRate-data}
\title{To be read by the software...}
\description{To be read by the software}
\details{To be read by the software
http://spreadsheets.google.com/pub?key=pk7kRzzfckbzz4AmH_e3DNA
Various sources
http://spreadsheets.google.com/pub?key=phAwcNAVuyj0NpF2PTov2Cw&gid=2
log

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmInfantMortalityRate}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=phAwcNAVuyj0NpF2PTov2Cw&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

