\name{gmMalariaReportedCases-data}
\alias{gmMalariaReportedCases-data}
\alias{gmMalariaReportedCases-data}
\title{Reported malaria cases...}
\description{Reported malaria cases}
\details{Reported malaria cases
Definition and explanations
Indicator name Reported malaria cases
Definition of indicator Total number of reported malaria cases during the given year.
Unit of measurement
Data source 
Source organization(s) World Health Organization
Link to source organization http://www.who.int
Complete reference Global Tuberculosis Database, WHO
Link to complete reference http://www.who.int/globalatlas/dataQuery/default.asp
Specific information about this indicator
Uploader Gapminder
[Add other fields as required]

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmMalariaReportedCases}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=pp59adS3CHWczfPHQMiqxCg&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

