\name{gmMaternalMortalityRate-data}
\alias{gmMaternalMortalityRate-data}
\alias{gmMaternalMortalityRate-data}
\title{Maternal mortality (per 100,000 live births)...}
\description{Maternal mortality (per 100,000 live births)}
\details{Maternal mortality (per 100,000 live births)
Definition and explanations
Indicator name Maternal mortality (per 100,000 live births)
Definition of indicator The number of maternal deaths divided by the number of live births in a given
year, multiplied by 100,000. Maternal death is defined as the death of a women while pregnant or within the 42
days after termination of that pregnancy, regardless of the length and site of the pregnancy, from a cause
related to or aggravated by the pregnancy.
Unit of measurement
Data source
Source organization(s) UN Statistics Division
Link to source organization http://data.un.org
Complete reference UN (United Nations). 2008. "UNData". Database.
Link to complete reference http://data.un.org/wiki/CDB.ashx
Specific information about this indicator
Uploader Gapminder
Extracted from database on 5/27/2008 16:53

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmMaternalMortalityRate}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=pyj6tScZqmEcVezxiMlWaRw&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

