\name{gmOilProduction-data}
\alias{gmOilProduction-data}
\alias{gmOilProduction-data}
\title{Oil Production total (tonnes per year)...}
\description{Oil Production total (tonnes per year)}
\details{Oil Production total (tonnes per year)
Definition and explanations
Indicator name Oil Production total (tonnes per year)
Definition of indicator Oil Production total (tonnes per year)
Unit of measurement
Data source 
Source organization(s) BP
Link to source organization http://www.bp.com/
Complete reference BP. 2009. Statistical Review of World Energy
Link to complete reference http://www.bp.com/statisticalreview
Specific information about this indicator
Uploader Gapminder
Time of uploading 2/1 2010

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmOilProduction}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=pyj6tScZqmEdNIa3ckVXaCQ&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

