\name{gmPeopleLivingWithHIV-data}
\alias{gmPeopleLivingWithHIV-data}
\alias{gmPeopleLivingWithHIV-data}
\title{People living with HIV/AIDS...}
\description{People living with HIV/AIDS}
\details{People living with HIV/AIDS
The data you see in this graph is mainly from UNAIDS, but Gapminder has complemented this dataset with
data and estimates directly from the countries.
The link below takes you to Gapminder's documentation
page, which contains the details on how the compilation was done and the sources for each observation. 

Definition and explanations
Indicator name People living with HIV/AIDS
Definition of indicator Number of people, all ages, living with HIV/AIDS
Unit of measurement Person
Data source 
Source organization(s) Various sources
Link to complete documentation
http://www.gapminder.org/gapminder-world/documentation/#gd006
Specific information about this indicator
Uploader Gapminder
Uploaded 1/27/2009

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmPeopleLivingWithHIV}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=pyj6tScZqmEe1GaiYJX2qGA&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

