\name{gmPerCapitaCO2Emissions-data}
\alias{gmPerCapitaCO2Emissions-data}
\alias{gmPerCapitaCO2Emissions-data}
\title{Per capita CO2 emissions (metric tons of carbon)...}
\description{Per capita CO2 emissions (metric tons of carbon)}
\details{Per capita CO2 emissions (metric tons of carbon)
Definition and explanations
Indicator name Per capita CO2 emissions (metric tons of carbon)
Definition of indicator Average CO2 emission in metric tons per person during the given year,
calculated from deviding the total CO2 with the total population of the countries.
Unit of measurement
Data source 
Source organization(s) CDIAC (Carbon Dioxide Information Analysis Center)
Link to source organization http://cdiac.ornl.gov/
Complete reference CDIAC (Carbon Dioxide Information Analysis Center) - nation.1751_2006.csv
Link to complete reference http://cdiac.ornl.gov/ftp/ndp030/CSV-FILES/
Specific information about this indicator
Uploader Gapminder

Calculation method Original data was only the amount of Carbon (C) and calculation has been done to
convert Carbon into Carbon Dioxide (CO2): (12+16*2)/12. Per capita emission data is based on calculations:
total CO2 emission / total population.
Total CO2 emission data from CDIAC, see this link for the original data.
http://spreadsheets.google.com/pub?key=phAwcNAVuyj1NHPC9MyZ9SQ
Total population data from Gapminder, see this link for the original data.
http://spreadsheets.google.com/pub?key=phAwcNAVuyj0XOoBL_n5tAQ
Only positive values have been included and displayed.

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmPerCapitaCO2Emissions}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=phAwcNAVuyj1gkNuUEXOGag&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

