\name{gmPlantedForestArea-data}
\alias{gmPlantedForestArea-data}
\alias{gmPlantedForestArea-data}
\title{Planted forest area (ha)...}
\description{Planted forest area (ha)}
\details{Planted forest area (ha)
Definition and explanations
Indicator name Planted forest area (ha)
Definition of indicator Total area of forest land (hectares) established through planting or
seeding, basically of introduced species and in some cases native species.
Unit of measurement ha
Data source 
Source organization(s) Food and Agriculture Organization (FAO)
Link to source organization http://www.fao.org/forestry/sofo/en/
Complete reference Complete reference
Link to complete reference Link to complete reference
Specific information about this indicator
Uploader Gapminder
[Add other fields as required]

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmPlantedForestArea}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=pp59adS3CHWc4aJd9fV8zZg&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

