\name{gmPrimaryEnergyConsumption-data}
\alias{gmPrimaryEnergyConsumption-data}
\alias{gmPrimaryEnergyConsumption-data}
\title{Primary energy consumption total (tonnes oil equivalent)...}
\description{Primary energy consumption total (tonnes oil equivalent)}
\details{Primary energy consumption total (tonnes oil equivalent)
Definition and explanations
Indicator name Primary energy consumption total (tonnes oil equivalent)
Definition of indicator Primary energy consumption total (tonnes oil equivalent)
Unit of measurement
Data source 
Source organization(s) BP
Link to source organization http://www.bp.com/
Complete reference BP. 2009. Statistical Review of World Energy
Link to complete reference http://www.bp.com/statisticalreview
Specific information about this indicator
Uploader Gapminder
Time of uploading 2/1 2010

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmPrimaryEnergyConsumption}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=pyj6tScZqmEeTCOezV8a3HA&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

