\name{gmPrimaryForestArea-data}
\alias{gmPrimaryForestArea-data}
\alias{gmPrimaryForestArea-data}
\title{Primary forest area (ha)...}
\description{Primary forest area (ha)}
\details{Primary forest area (ha)
Definition and explanations
Indicator name Primary forest area (ha)
Definition of indicator Total area of forest land (hectares) with native species, where there are no
clearly visible indications of human activities and the ecological processes are not significantly disturbed.
Unit of measurement ha
Data source 
Source organization(s) Food and Agriculture Organization (FAO)
Link to source organization http://www.fao.org/forestry/sofo/en/
Complete reference Complete reference
Link to complete reference Link to complete reference
Specific information about this indicator
Uploader Gapminder
[Add other fields as required]

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmPrimaryForestArea}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=pp59adS3CHWeECA6Gf__BNQ&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

