\name{gmSugarConsumption-data}
\alias{gmSugarConsumption-data}
\alias{gmSugarConsumption-data}
\title{Sugar per person (g per day)...}
\description{Sugar per person (g per day)}
\details{Sugar per person (g per day)
Definition and explanations
Indicator name Sugar per person (g per day)
Definition of indicator The food consumption quantity (grams per person and day) of sugar and sweeters
Unit of measurement Grams per person and day
Data source
Sources FAO stats
Link to source organizations
http://faostat.fao.org/site/609/DesktopDefault.aspx?PageID=609#cible
Downloaded 10/1/2008
Notes data for 2004 based on a rough extrapolation
some other country notes, see fottnotes

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmSugarConsumption}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=phAwcNAVuyj2sdmdhX9zuKg&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

