\name{gmSulfurEmissions-data}
\alias{gmSulfurEmissions-data}
\alias{gmSulfurEmissions-data}
\title{Total sulfur emission (kilotonnes/year)...}
\description{Total sulfur emission (kilotonnes/year)}
\details{Total sulfur emission (kilotonnes/year)
The data is based on an article by David Stern (see below). In some cases adjustments had to made for new
countries emerging from former countries (e.g. the USSR).
Definition of indicator
Definition Human emission of sulfur.
Unit Kilotonnes (milion kilograms) per year.
Sources
The data is based on the following article: Stern D. I. (2006); "Reversal in the trend of global anthropogenic sulfur emissions",
Global Environmental Change 16, 207-220.
The article is avilable at: http://www.sterndavidi.com/Publications/GEC2006.pdf
The data we used is available at: http://www.sterndavidi.com/datasite.html
Data adjustments by Gapminder
A number of countries in the dataset were created when former countries were split. For the earlier
years Stern only provide data for the totals of these fomer countries. For later years data for the individual
new countries were available.
We assumed that the new country's had a constant share in the total of the former country, whenever data
was missing. The share used was normally the same as for the earliest year with data for the individual
country. Sometimes we used a rough average of the "typical share" during the years with data. The shares always add
up to one for all the members of each particular fomrer country.
See the "footnotes" tab for information on the adjustments we made for each individual country. 
Uploaded
2010: januari 26

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmSulfurEmissions}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=t9SYWh7siLJDzyZYN1R4HfQ&gid=2&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

