\name{gmTotalForestArea-data}
\alias{gmTotalForestArea-data}
\alias{gmTotalForestArea-data}
\title{Forest land, total area (ha)...}
\description{Forest land, total area (ha)}
\details{Forest land, total area (ha)
Definition and explanations
Indicator name Forest land, total area (ha)
Definition of indicator Total area of land (hectares) which is spanning more than 0.5 hectares with
trees higher than 5 meters and a canopy cover of more than 10\%, or trees able to reach these thresholds in situ. It
does not include land that is predominantly under agricultural or urban land use.
Unit of measurement ha
Data source 
Source organization(s) Food and Agriculture Organization (FAO)
Link to source organization http://www.fao.org/forestry/sofo/en/
Complete reference Complete reference
Link to complete reference Link to complete reference
Specific information about this indicator
Uploader Gapminder
[Add other fields as required]

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmTotalForestArea}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=pp59adS3CHWeB1N1HlpFQVQ&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

