\name{gmUnder5MortalityRate-data}
\alias{gmUnder5MortalityRate-data}
\alias{gmUnder5MortalityRate-data}
\title{Under-five mortality rate (per 1,000 live births)...}
\description{Under-five mortality rate (per 1,000 live births)}
\details{Under-five mortality rate (per 1,000 live births)
Definition and explanations
Indicator name Under-five mortality rate (per 1,000 live births)
Definition of indicator The probability that children born in a specific year will die before reaching
the age of five, if the age-specific mortality rates remain the same. Expressed as a rate per 1,000 live
births. From CME Info, Child Mortality Estimates by the Inter-agency Group for Child Mortality Estimation,
which includes technical experts from WHO, UNICEF, the United Nations Population Division, the World Bank,
and independent experts.
Unit of measurement Death of children under five years of age per 1,000 live births
Data source
Source organization(s) Inter-agency Group for Child Mortality Estimation (IGME)
Link to source organization
Complete reference www.childmortality.org
Links to complete reference
Specific information about this indicator
Uploader Gapminder
[Add other fields as required] 1/28/2010

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmUnder5MortalityRate}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=phAwcNAVuyj05ZR69usyQIg&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

