\name{gmWomensAgeAtFirstMarriage-data}
\alias{gmWomensAgeAtFirstMarriage-data}
\alias{gmWomensAgeAtFirstMarriage-data}
\title{Mean age at 1st marriage of women...}
\description{Mean age at 1st marriage of women}
\details{Mean age at 1st marriage of women
Gapminder has compiled the data you see in this graph from several sources, such as official
international statistics, various historical sources and our own estimates. 
The link below takes you to
Gapminder's documentation page, which contains the details on how the compilation was done and the sources for each
observation. 

Definition and explanations
Indicator name Mean age at 1st marriage of women
Definition of indicator The average age, in years, of first marriage for women. Women who never married
are excluded. Cohabitation is excluded. The data are based on multple sources and definitions might vary.
WE DISCOURAGE THE USE OF THIS DATASET FOR STATISTICAL ANALYSIS
PLEASE CONSULT THE FULL DOCUMENTATION FOR MORE DETAILS
Link to full documentation of this indicator 
http://www.gapminder.org/downloads/documentation/#gd009
Specific information about this indicator
Uploader Gapminder, Mattias Lindgren
Data uploaded 10/29/2009

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmWomensAgeAtFirstMarriage}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=t4eF8H_jq_xyKCUHAX6VT1g&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

