\name{gmWoodRemoval-data}
\alias{gmWoodRemoval-data}
\alias{gmWoodRemoval-data}
\title{Wood removal (cubic meter)...}
\description{Wood removal (cubic meter)}
\details{Wood removal (cubic meter)
Definition and explanations
Indicator name Wood removal (cubic meter)
Definition of indicator The wood removed (volume of round wood over bark) for production of goods and
services other than energy production (wood fuel), excluding felled trees left in the forest.
Unit of measurement (m3)
Data source 
Source organization(s) Food and Agriculture Organization (FAO)
Link to source organization http://www.fao.org/forestry/sofo/en/
Complete reference Complete reference
Link to complete reference Link to complete reference
Specific information about this indicator
Uploader Gapminder
[Add other fields as required]

Please note that neither Gapminder nor the package developer/maintainer are the data provider, except for a few cases.
Therefore you will have to go to the source to find out the terms of use for the specific indicator.}
\alias{gmWoodRemoval}
\references{\url{http://www.gapminder.org}
\url{http://spreadsheets.google.com/pub?key=pp59adS3CHWe8O-N9RgxzDw&gid=1&output=csv}}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

