\name{mediawiki}
\alias{mediawiki}
\title{Creates a mediawiki S3 object.}
\usage{mediawiki(api="")
}
\description{Creates a mediawiki S3 object.}
\details{Set the api url as parameter.}
\value{an object of class "mediawiki"}
\references{\url{http://www.mediawiki.org/wiki/API}}
\author{Karsten Weinert \email{k.weinert@gmx.net}}
\arguments{\item{api}{string pointing to the API url. Usually ends with api.php}
}

