\name{print.ckan}
\alias{print.ckan}
\title{String representation of a ckan object...}
\usage{print.ckan(obj, ...)
}
\description{String representation of a ckan object}
\details{Displays the API url.}
\value{string}
\seealso{\code{ckan} for a description of the datasource}
\author{Karsten Weinert \email{k.weinert@gmx.net}}
\arguments{\item{obj}{reference of the ckan object}
}

