\name{print.dbpedia}
\alias{print.dbpedia}
\title{String representation of a dbpedia object...}
\usage{print.dbpedia(obj, ...)
}
\description{String representation of a dbpedia object}
\details{Displays the names of available entities. For documentation of a specific dataset see the wzddata package.}
\value{string}
\seealso{\code{dbpedia} for a description of the datasource}
\author{Karsten Weinert \email{k.weinert@gmx.net}}
\arguments{\item{obj}{reference of the dbpedia object}
}

