\name{print.gapminder}
\alias{print.gapminder}
\title{Print information on a gapminder S3 object...}
\usage{print.gapminder(x, ...)
}
\description{Print information on a gapminder S3 object}
\details{Displays the names of available entities. For documentation of a specific dataset see the wzddata package.}
\value{used for its side effect}
\seealso{\code{gapminder} for a description of the datasource}
\author{Karsten Weinert \email{k.weinert@gmx.net}}
\arguments{\item{x}{reference of the gapminder object}
\item{...}{unused (from generic)}
}

