\name{print.mediawiki}
\alias{print.mediawiki}
\title{Prints information on a mediawiki API object.}
\usage{print.mediawiki(x, ...)
}
\description{Prints information on a mediawiki API object.}
\value{used for its side effect.}
\author{Karsten Weinert \email{k.weinert@gmx.net}}
\arguments{\item{x}{mediawiki S3 object}
\item{...}{unused (from generic)}
}

