\name{query}
\alias{query}
\title{Generic S3 method to query different data sources.}
\usage{query(x, ...)
}
\description{Generic S3 method to query different data sources.}
\details{CKAN describes itself as a registry of open data and content packages. CKAN provides ways to find, download and share open content and data.}
\value{depends on the data source}
\seealso{\code{query.mediawiki} for querying mediawiki}
\seealso{\code{query.gapminder} for querying gapminder datasets}
\seealso{\code{query.ckan} for querying ckan data entities}
\author{Karsten Weinert \email{k.weinert@gmx.net}}
\arguments{\item{x}{data source object}
}

