\name{RweaveMWiki}
\alias{RweaveMWiki}
\title{Driver for Sweave() which outputs MWiki formatted text.}
\description{Driver for Sweave() which outputs MWiki formatted text.}
\details{The code is, with only very few adaptions, taken from the package ascii by David Hajage.}
\value{a list object suitable as driver parameter for Sweave.}
\references{\url{http://wagezudenken.blogspot.com/2009/12/r-workflow-test.html}}
\author{Karsten Weinert \email{k.weinert@gmx.net}}
\arguments{\item{None}{}}
