\name{RweaveMWikiFinish}
\alias{RweaveMWikiFinish}
\title{Internal function for RweaveMWiki.}
\usage{RweaveMWikiFinish(object, error=FALSE)RweaveMWikiFinish(object, error=FALSE)
RweaveMWikiFinish(object, error=FALSE)
RweaveMWikiFinish(object, error=FALSE)
RweaveMWikiFinish(object, error=FALSE)
RweaveMWikiFinish(object, error=FALSE)
RweaveMWikiFinish(object, error=FALSE)
}
\description{Internal function for RweaveMWiki.}
\details{Hook called at the end of a Sweave call.\code{RweaveMWikiFinish}: Hook called at the end of a Sweave call.


\code{RweaveMWikiFinish}: Hook called at the end of a Sweave call.


\code{RweaveMWikiFinish}: Hook called at the end of a Sweave call.


\code{RweaveMWikiFinish}: Hook called at the end of a Sweave call.


\code{RweaveMWikiFinish}: Hook called at the end of a Sweave call.


\code{RweaveMWikiFinish}: Hook called at the end of a Sweave call.


}
\value{object\code{RweaveMWikiFinish}: object


\code{RweaveMWikiFinish}: object


\code{RweaveMWikiFinish}: object


\code{RweaveMWikiFinish}: object


\code{RweaveMWikiFinish}: object


\code{RweaveMWikiFinish}: object


}
\arguments{\item{object}{the object}
\item{error}{default FALSE}}
\alias{RweaveMWikiFinish}

\alias{RweaveMWikiFinish}

\alias{RweaveMWikiFinish}

\alias{RweaveMWikiFinish}

\alias{RweaveMWikiFinish}

\alias{RweaveMWikiFinish}

