\name{RweaveMWikiOptions}
\alias{RweaveMWikiOptions}
\title{Internal function for RweaveMWiki.}
\usage{RweaveMWikiOptions(options)RweaveMWikiOptions(options)
RweaveMWikiOptions(options)
RweaveMWikiOptions(options)
RweaveMWikiOptions(options)
RweaveMWikiOptions(options)
RweaveMWikiOptions(options)
}
\description{Internal function for RweaveMWiki.}
\details{Check the sanity of the options. Stops on invalid options.\code{RweaveMWikiOptions}: Check the sanity of the options. Stops on invalid options.


\code{RweaveMWikiOptions}: Check the sanity of the options. Stops on invalid options.


\code{RweaveMWikiOptions}: Check the sanity of the options. Stops on invalid options.


\code{RweaveMWikiOptions}: Check the sanity of the options. Stops on invalid options.


\code{RweaveMWikiOptions}: Check the sanity of the options. Stops on invalid options.


\code{RweaveMWikiOptions}: Check the sanity of the options. Stops on invalid options.


}
\value{options\code{RweaveMWikiOptions}: options


\code{RweaveMWikiOptions}: options


\code{RweaveMWikiOptions}: options


\code{RweaveMWikiOptions}: options


\code{RweaveMWikiOptions}: options


\code{RweaveMWikiOptions}: options


}
\arguments{\item{options}{the options}

}
\alias{RweaveMWikiOptions}

\alias{RweaveMWikiOptions}

\alias{RweaveMWikiOptions}

\alias{RweaveMWikiOptions}

\alias{RweaveMWikiOptions}

\alias{RweaveMWikiOptions}

