\name{RweaveMWikiRuncode}
\alias{RweaveMWikiRuncode}
\title{Internal function for RweaveMWiki.}
\usage{RweaveMWikiRuncode(object, chunk, options)RweaveMWikiRuncode(object, chunk, options)
RweaveMWikiRuncode(object, chunk, options)
RweaveMWikiRuncode(object, chunk, options)
RweaveMWikiRuncode(object, chunk, options)
RweaveMWikiRuncode(object, chunk, options)
RweaveMWikiRuncode(object, chunk, options)
}
\description{Internal function for RweaveMWiki.}
\details{runs one chunk of R code\code{RweaveMWikiRuncode}: runs one chunk of R code


\code{RweaveMWikiRuncode}: runs one chunk of R code


\code{RweaveMWikiRuncode}: runs one chunk of R code


\code{RweaveMWikiRuncode}: runs one chunk of R code


\code{RweaveMWikiRuncode}: runs one chunk of R code


\code{RweaveMWikiRuncode}: runs one chunk of R code


}
\value{object\code{RweaveMWikiRuncode}: object


\code{RweaveMWikiRuncode}: object


\code{RweaveMWikiRuncode}: object


\code{RweaveMWikiRuncode}: object


\code{RweaveMWikiRuncode}: object


\code{RweaveMWikiRuncode}: object


}
\arguments{\item{object}{the object}
\item{chunk}{the chunk}
\item{options}{the options}}
\alias{RweaveMWikiRuncode}

\alias{RweaveMWikiRuncode}

\alias{RweaveMWikiRuncode}

\alias{RweaveMWikiRuncode}

\alias{RweaveMWikiRuncode}

\alias{RweaveMWikiRuncode}

