\name{RweaveMWikiSetup}
\alias{RweaveMWikiSetup}
\title{Internal function for RweaveMWiki.}
\usage{RweaveMWikiSetup(file, syntax, output, quiet=FALSE, debug=FALSE, ...)}
\description{Internal function for RweaveMWiki.}
\details{Sets the options for the RweaveMWiki driver.}
\value{a list with components output, debug, quiet, syntax, options, chunkout, srclines and srcfile}
\arguments{\item{file}{input file name}
\item{syntax}{syntax object}
\item{output}{output file name (NULL)}
\item{quiet}{print some information (FALSE)}
\item{debug}{print more information (FALSE)}
\item{...}{prefix, prefix.string, engine, print, eval, fig, ext, png, jpg, pdf, eps, width, height, term, echo, keep.source, results, split, strip.white, include, pdf.version, pdf.encoding, concordance, expand}}
