\name{RweaveMWikiWritedoc}
\alias{RweaveMWikiWritedoc}
\title{Internal function for RweaveMWiki.}
\usage{RweaveMWikiWritedoc(object, chunk)RweaveMWikiWritedoc(object, chunk)
RweaveMWikiWritedoc(object, chunk)
RweaveMWikiWritedoc(object, chunk)
RweaveMWikiWritedoc(object, chunk)
RweaveMWikiWritedoc(object, chunk)
RweaveMWikiWritedoc(object, chunk)
}
\description{Internal function for RweaveMWiki.}
\details{writes out a evaluated code chunk.\code{RweaveMWikiWritedoc}: writes out a evaluated code chunk.


\code{RweaveMWikiWritedoc}: writes out a evaluated code chunk.


\code{RweaveMWikiWritedoc}: writes out a evaluated code chunk.


\code{RweaveMWikiWritedoc}: writes out a evaluated code chunk.


\code{RweaveMWikiWritedoc}: writes out a evaluated code chunk.


\code{RweaveMWikiWritedoc}: writes out a evaluated code chunk.


}
\value{object\code{RweaveMWikiWritedoc}: object


\code{RweaveMWikiWritedoc}: object


\code{RweaveMWikiWritedoc}: object


\code{RweaveMWikiWritedoc}: object


\code{RweaveMWikiWritedoc}: object


\code{RweaveMWikiWritedoc}: object


}
\arguments{\item{object}{the object}
\item{chunk}{the chunk}}
\alias{RweaveMWikiWritedoc}

\alias{RweaveMWikiWritedoc}

\alias{RweaveMWikiWritedoc}

\alias{RweaveMWikiWritedoc}

\alias{RweaveMWikiWritedoc}

\alias{RweaveMWikiWritedoc}

