\name{wzd-package}
\alias{wzd-package}
\alias{wzd-package}
\title{Accompanying package to the blog wagezudenken.blogspot.com}
\description{Accompanying package to the blog wagezudenken.blogspot.com}
\details{The package features datasets from various topics, data visualizations including cartography, statistical methods and tools; as 
discussed on the blog.}
\alias{wzd}
\docType{package}
\author{Karsten Weinert \email{k.weinert@gmx.net}}
\references{\url{http://wagezudenken.blogspot.com}}
\keyword{package}

