\name{ckan}
\alias{ckan}
\title{Creates a CKAN data source (S3 object).}
\usage{ckan(api="http://ckan.net")
}
\description{Creates a CKAN data source (S3 object).}
\details{CKAN describes itself as a registry of open data and content packages. CKAN provides ways to find, download and share open content and data.}
\value{an object of class "ckan"}
\references{\url{http://knowledgeforge.net/ckan/doc/ckan/api.html}}
\author{Karsten Weinert \email{k.weinert@gmx.net}}
\arguments{\item{api}{string pointing to the api url as parameter. Default is http://ckan.net}
}

