\name{print.ckan}
\alias{print.ckan}
\title{Print information on a ckan S3 object...}
\usage{print.ckan(x, ...)
}
\description{Print information on a ckan S3 object}
\details{Displays the API url.}
\value{used for its side effect}
\seealso{\code{ckan} for a description of the datasource}
\author{Karsten Weinert \email{k.weinert@gmx.net}}
\arguments{\item{x}{reference of the ckan object}
\item{...}{unused (from generic)}
}

