\name{print.gapminder}
\alias{print.gapminder}
\title{String representation of a gapminder object...}
\usage{print.gapminder(obj, ...)
}
\description{String representation of a gapminder object}
\details{Displays the names of available entities. For documentation of a specific dataset see the wzddata package.}
\value{string}
\seealso{\code{gapminder} for a description of the datasource}
\author{Karsten Weinert \email{k.weinert@gmx.net}}
\arguments{\item{obj}{reference of the gapminder object}
}

