\name{query}
\alias{query}
\title{Generic S3 method to query different data sources.}
\usage{query(obj, ...)
}
\description{Generic S3 method to query different data sources.}
\details{This method is implemented for various datasources of this package, see below.}
\value{depends on the data source}
\seealso{\code{query.mediawiki} for querying mediawiki}
\seealso{\code{query.gapminder} for querying gapminder datasets}
\seealso{\code{query.ckan} for querying ckan data entities}
\author{Karsten Weinert \email{k.weinert@gmx.net}}
\arguments{\item{obj}{data source object}
\item{...}{no default use}
}

