\name{RweaveEvalWithOpt}
\alias{RweaveEvalWithOpt}
\title{Internal function for RweaveMWiki.}
\usage{RweaveEvalWithOpt(expr, options)}
\description{Internal function for RweaveMWiki.}
\details{evaluates an R expression}
\value{result of eval(expr)}
\arguments{\item{expr}{R expression as string}
\item{options}{list of options. Components print, eval and term are read}}
