\name{RweaveMWikiOptions}
\alias{RweaveMWikiOptions}
\title{Internal function for RweaveMWiki.}
\usage{RweaveMWikiOptions(options)
}
\description{Internal function for RweaveMWiki.}
\details{Check the sanity of the options. Stops on invalid options.}
\value{options}
\arguments{\item{options}{the options}
}

