\name{RweaveMWikiRuncode}
\alias{RweaveMWikiRuncode}
\title{Internal function for RweaveMWiki.}
\usage{RweaveMWikiRuncode(object, chunk, options)}
\description{Internal function for RweaveMWiki.}
\details{runs one chunk of R code}
\value{object}
\arguments{\item{object}{the object}
\item{chunk}{the chunk}
\item{options}{the options}}
