\name{RweaveMWikiWritedoc}
\alias{RweaveMWikiWritedoc}
\title{Internal function for RweaveMWiki.}
\usage{RweaveMWikiWritedoc(object, chunk)}
\description{Internal function for RweaveMWiki.}
\details{writes out a evaluated code chunk.}
\value{object}
\arguments{\item{object}{the object}
\item{chunk}{the chunk}}
