\name{worldSimple-data}
\alias{worldSimple-data}
\alias{worldSimple-data}
\title{Spatial dataset for world maps.}
\description{Spatial dataset for world maps.}
\details{The data is taken from Bjorn Sandvik's site www.thematicmapping.org and processed using Roger Bivand's instructions on 
spatial.nhh.no/R/etc/world_map_intro.html.

Bjorn Sandvik's data is protected by a CC-SA license.}
\alias{worldSimple}
\docType{data}
\author{Karsten Weinert \email{k.weinert@gmx.net}}

