\name{wzd-package}
\title{Accompanying package to the blog wagezudenken.blogspot.com}
\description{Accompanying package to the blog wagezudenken.blogspot.com}
\details{\tabular{ll}{
Package: \tab wzd\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-02-22\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}

The package features datasets from various topics, data visualizations including cartography, statistical methods and tools; as 
discussed on the blog.\code{wzd-package}: \tabular{ll}{
Package: \tab wzd\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-02-22\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}

The package features datasets from various topics, data visualizations including cartography, statistical methods and tools; as 
discussed on the blog.


\code{wzd-package}: \tabular{ll}{
Package: \tab wzd\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-02-22\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}

The package features datasets from various topics, data visualizations including cartography, statistical methods and tools; as 
discussed on the blog.


\code{wzd-package}: \tabular{ll}{
Package: \tab wzd\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-02-22\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}

The package features datasets from various topics, data visualizations including cartography, statistical methods and tools; as 
discussed on the blog.


\code{wzd-package}: \tabular{ll}{
Package: \tab wzd\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-02-22\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}

The package features datasets from various topics, data visualizations including cartography, statistical methods and tools; as 
discussed on the blog.


\code{wzd-package}: \tabular{ll}{
Package: \tab wzd\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-02-22\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}

The package features datasets from various topics, data visualizations including cartography, statistical methods and tools; as 
discussed on the blog.


\code{wzd-package}: \tabular{ll}{
Package: \tab wzd\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-02-22\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}

The package features datasets from various topics, data visualizations including cartography, statistical methods and tools; as 
discussed on the blog.


}
\alias{wzd}
\docType{packagepackagepackagepackagepackagepackagepackage}
\author{Karsten Weinert \email{k.weinert@gmx.net}}
\references{\url{http://wagezudenken.blogspot.com}\url{http://wagezudenken.blogspot.com}\url{http://wagezudenken.blogspot.com}\url{http://wagezudenken.blogspot.com}\url{http://wagezudenken.blogspot.com}\url{http://wagezudenken.blogspot.com}\url{http://wagezudenken.blogspot.com}}
\keyword{package}
\alias{wzd-package}
\alias{wzd-package}
\alias{wzd}
\keyword{package}

\alias{wzd-package}
\alias{wzd-package}
\alias{wzd}
\keyword{package}

\alias{wzd-package}
\alias{wzd-package}
\alias{wzd}
\keyword{package}

\alias{wzd-package}
\alias{wzd-package}
\alias{wzd}
\keyword{package}

\alias{wzd-package}
\alias{wzd-package}
\alias{wzd}
\keyword{package}

\alias{wzd-package}
\alias{wzd-package}
\alias{wzd}
\keyword{package}

